Added cocoapods support and updated documentation.
